Wave 3.2: rename entity_wire.rs to store/wire.rs
3.1 landed the same EntityRow::into_wire split under store/wire.rs, for
the identical reason (tests/one_predicate.rs false-positives on entity.rs).
Matching the path means an integration merge conflicts textually in one
file with an obvious resolution, rather than compiling two inherent
into_wire methods and failing at E0592 with nothing in the diff pointing
at why.
