+++ crates/altaird/src/store/mod.rs
@@ -6,10 +6,10 @@
 
 pub mod audience;
 pub mod entity;
-mod entity_wire;
 pub mod ids;
 pub mod relation;
 pub mod tx;
+mod wire;
 
 pub use audience::{Bind, CandidateQuery, LifecycleScope, ReadScope, WriteScope};
 pub use ids::{EntityId, MemberId};
